Moved SimpleSearch AsJsonValue to separate file
